pyproject.toml
[project]
name = "social-media-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.106.1",
    "openai-agents>=0.2.11",
    "python-dotenv>=1.1.1",
    "pytube>=15.0.0",
    "streamlit>=1.49.1",
    "youtube-transcript-api>=1.2.2",
]
